fix(pipeline): 3 compile errors from run 34920556129
CI-log-verified (win job 104227513026):
- E0308 background.rs:208/211 — k-means centroid sums are u32, cast to u8
- E0599 background.rs:344 — RasterView trait (luma_row) not in scope

+++ crates/isg-native/src/pipeline/background.rs
@@ -5,7 +5,7 @@
 //! are RGB outliers but perceptually identical to the background, so a ΔE
 //! threshold cuts speckle sharply without touching real icons.
 
-use isg_core::ForegroundMask;
+use isg_core::{ForegroundMask, RasterView};
 
 use super::raster::SheetRaster;
 
@@ -205,10 +205,18 @@ fn detect_kmeans_corners(r: &SheetRaster) -> Option<BackgroundModel> {
             }
         }
         if n0 > 0 {
-            c0 = [s0[0] / n0, s0[1] / n0, s0[2] / n0];
+            c0 = [
+                (s0[0] / n0) as u8,
+                (s0[1] / n0) as u8,
+                (s0[2] / n0) as u8,
+            ];
         }
         if n1 > 0 {
-            c1 = [s1[0] / n1, s1[1] / n1, s1[2] / n1];
+            c1 = [
+                (s1[0] / n1) as u8,
+                (s1[1] / n1) as u8,
+                (s1[2] / n1) as u8,
+            ];
         }
     }
     // Border ownership vote → background cluster.
